% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Run the Branch Mover App}
\usage{
app(username = NULL, include_fork = FALSE, include_private = TRUE)
}
\arguments{
\item{username}{The user name or organization name whose repos you would like
to move. If the username is \code{NULL} or the same as the authenticated user,
you can use branchMover to change the default branch of private repos.
Otherwise, the GitHub API doesn't allow us to list private repos for
organizations (or, if they do, please let me know!).}

\item{include_fork, include_private}{Should forks (default \code{FALSE}) or private
repos (default \code{TRUE}) be included?}
}
\value{
Runs a Shiny app. Your settings and preferences are remembered
between sessions using RStudio user preferences.
}
\description{
Run the Branch Mover App
}
\details{
The Branch Mover app helps you move the default branches of your repositories
on GitHub. For more details about how this works or why you'd want to use
Branch Mover, please see the blog post:
\url{https://www.garrickadenbuie.com/blog/branchmover/}

The app uses the \href{https://usethis.r-lib.org/}{usethis} and
\href{https://gh.r-lib.org/}{gh} packages. You need to configure gh with a
\href{https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token}{Personal Access Token} to be able to authenticate with the GitHub API. Read more about setting up a PAT in one of these places:
\itemize{
\item \href{https://usethis.r-lib.org/articles/articles/git-credentials.html}{Managing Git(Hub) Credentials}
\emph{usethis article}
\item \href{https://happygitwithr.com/https-pat.html}{Personal access token for HTTPS}
\emph{guidance from \href{https://happygitwithr.com/}{Happy Git and GitHub for the useR}}
\item \href{https://gh.r-lib.org/articles/managing-personal-access-tokens.html}{Managing Personal Access Tokens}
\emph{gh article}
}
}
