% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchMover-package.R
\docType{package}
\name{branchMover-package}
\alias{branchMover}
\alias{branchMover-package}
\title{branchMover: An RStudio Addin to Help Move the Default Branch of Your GitHub Repos}
\description{
Helper app to move the default branch of GitHub repositories under your control.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/branchMover}
  \item Report bugs at \url{https://github.com/gadenbuie/branchMover/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
