% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchMover-package.R
\docType{package}
\name{branchMover-package}
\alias{branchMover}
\alias{branchMover-package}
\title{branchMover: An RStudio Addin to Help Move the Default Branch of Your GitHub Repos}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
